+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to the "Diff Folders" extension will be documented in this f
 
 ### Added
 - Added drag'n drop support for path names in the list view which makes cross diffs possible. Please read the description `List Drag'n Drop` to see how it works.
-- Added groups to favorites to group your favorite diff and open all at once.
+- Added groups for favorites to group your diffs and open all at once.
 - Added `l13Diff.initialFavoriteGroupState` to set the initial state of the groups.
 
 ### Changed